System role for frontend
A role-based system for storing data in Pinia, restricting access based on role to different endpoints/pages

+++ UniversityProject/Frontend/frontend/src/api/Student.ts
@@ -1,11 +1,18 @@
 ﻿import api from "@/api/Api.ts";
 import {GetCookie} from "@/Function/CookiesFunction.ts";
+import {userAccessPage} from "@/stores/AccessPage.ts";
+import {ElMessage} from "element-plus";
 
 export const StudentResponse =
     {
         getStudents(sortKey?: any, sortType?: any, numberPage?: any, filter?: any, count?: any)
         {
             let token = GetCookie("accessJWT");
+            if(!userAccessPage().canAccessForAllOperationName("StudentRegister", ["Read", "All"]))
+            {
+                ElMessage.error('У вас нет доступа к данному действию');
+                return;
+            }
             return api.get(`Student?sortKey=${sortKey}&sortOrder=${sortType}&page=${numberPage}&count=${count}&filter=${filter}`,{
                 headers: {
                     'Authorization': token
@@ -22,6 +29,11 @@ export const StudentResponse =
         getStudent(id?: any)
         {
             let token = GetCookie("accessJWT");
+            if(!userAccessPage().canAccessForAllOperationName("StudentPage", ["Read", "All"]))
+            {
+                ElMessage.error('У вас нет доступа к данному действию');
+                return;
+            }
             return api.get(`Student/${id}`,{
                 headers: {
                     'Authorization': token
@@ -30,6 +42,11 @@ export const StudentResponse =
         deleteStudent(id?: any)
         {
             let token = GetCookie("accessJWT");
+            if(!userAccessPage().canAccessForAllOperationName("StudentPage", ["Delete", "All"]))
+            {
+                ElMessage.error('У вас нет доступа к данному действию');
+                return;
+            }
             return api.delete(`Student/${id}`,{
                 headers: {
                     'Authorization': token
@@ -38,6 +55,11 @@ export const StudentResponse =
         putStudent(id?: any, student?: any)
         {
             let token = GetCookie("accessJWT");
+            if(!userAccessPage().canAccessForAllOperationName("StudentPage", ["Update", "All"]))
+            {
+                ElMessage.error('У вас нет доступа к данному действию');
+                return;
+            }
             return api.put(`Student/${id}`, student,{
                 headers: {
                     'Authorization': token
@@ -46,6 +68,11 @@ export const StudentResponse =
         postStudent(student?: any)
         {
             let token = GetCookie("accessJWT");
+            if(!userAccessPage().canAccessForAllOperationName("StudentPage", ["Create", "All"]))
+            {
+                ElMessage.error('У вас нет доступа к данному действию');
+                return;
+            }
             return api.post(`Student`, student,{
                 headers: {
                     'Authorization': token
